Big Changes All around 8/6
modal practice added

buttons created, div created, button appended. Modal only shows when clicked. Adjusted the code to listen for clicks on the modal as well as all forms of exit on the modal if you want to use escape key, click out or hit close they all work. Pokemon are loaded with images, height and types into a modal that appears a listens on a click. All necessary code to create style modal and fill the modal with information from the Pokemon API and fetching the name, type, image from each of the pokemon to display in a certain manor when the button is clicked